[REFACTOR] Moved sendCmcdDataReport from CmcdModel to CmcdController
